Release: 0.3.0 - Added a "Plotting" tab that can be used to plot multiple data sets, test results, fit results, and simulation results in a single figure. Currently supports Nyquist, Bode (magnitude), and Bode (phase) plot types. - Added a setting for specifying the number of points per decade to use when drawing simulated responses as lines. - Added an "Edit" menu to the menu bar with "Undo" and "Redo" actions. - Added indicators for when the program is busy loading data files or loading/saving projects. - Added support for messages and progress bars to the indicator that is shown when the program is busy. - Added a check to make sure there are no UUID collisions when saving a project. - Added more keybindings. - Added more tooltips. - Updated the project file and config file structures to support the new features. - Updated keybinding documentation. - Updated existing tooltips. - Updated the labels of some buttons. - Updated the implementation of the recent projects table to update when a project is saved. - Updated how paths, which are passed as CLI arguments, to data files and projects, and project snapshots are processed when the program is started. - Updated CDC inputs in the fitting and simulation tabs to not require the Enter/Return key to be pressed before a fit or simulation can be performed. - Updated the confirmation window, which is shown when possibly saving over an existing project file, to automatically adjust its height to accommodate long paths. - Removed the restriction imposed on moving the error message window. - Fixed a bug that allowed duplicate entries in the list of recent projects. - Fixed a bug that allowed multiple instances of a project to be loaded. - Fixed a bug that triggered state snapshots twice when toggling the mask state of multiple points. - Fixed a bug that allowed specifying an invalid number of RC circuits for the Kramers-Kronig test.

+++ src/deareis/plot/shared.py
@@ -77,7 +77,13 @@ def auto_limits(axes: List[int]):
 
 def modal_window(label: str = "") -> int:
     assert type(label) is str
-    window: int = dpg.add_window(label=label, modal=True)
+    window: int = dpg.generate_uuid()
+    dpg.add_window(
+        label=label,
+        modal=True,
+        on_close=lambda: dpg.delete_item(window),
+        tag=window,
+    )
     x: int
     y: int
     w: int
@@ -115,8 +121,10 @@ class Plot:
         window: int = dpg.get_item_parent(plot)
 
         def close_window_callback():
-            dpg.hide_item(window)
-            dpg.delete_item(handler)
+            if dpg.does_item_exist(window):
+                dpg.hide_item(window)
+            if dpg.does_item_exist(handler):
+                dpg.delete_item(handler)
 
         with dpg.handler_registry(tag=handler):
             dpg.add_key_release_handler(
